+++ CHANGELOG.md
@@ -10,6 +10,9 @@ subsampling:
   1.1.0-alpha02. [#31](https://github.com/panpf/zoomimage/issues/31)
 * improve: GlideModeToImageSource and PicassoDataToImageSource are now priority for user
   registration
+* improve: Improve BitmapRegionDecoderDecodeHelper. It no longer closes the input stream. It will
+  now close the input stream when BitmapRegionDecoderDecodeHelper is
+  destroyed. [#29](https://github.com/panpf/zoomimage/issues/29)
 
 ## 1.1.0-alpha02